Ecommerce-API/pyproject.toml
[project]
name = "ecommerce-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10, <3.12"
dependencies = [
    "alembic==1.12.1",
    "annotated-types==0.6.0",
    "anyio==3.7.1",
    "bcrypt==4.0.1",
    "click==8.1.7",
    "colorama==0.4.6",
    "dnspython==2.4.2",
    "ecdsa==0.18.0",
    "email-validator==2.1.0.post1",
    "exceptiongroup==1.2.0",
    "fastapi==0.104.1",
    "fastembed==0.2.0",
    "greenlet==3.0.1",
    "h11==0.14.0",
    "httptools==0.6.1",
    "httpx==0.24.1",
    "idna==3.4",
    "mako==1.3.0",
    "markupsafe==2.1.3",
    "neo4j>=6.1.0",
    "passlib==1.7.4",
    "pillow>=12.1.0",
    "psycopg2-binary==2.9.9",
    "pyasn1==0.5.1",
    "pydantic==2.5.1",
    "pydantic-core==2.14.3",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.0",
    "python-jose==3.3.0",
    "python-multipart==0.0.6",
    "pyyaml==6.0.1",
    "qdrant-client==1.12.0",
    "rsa==4.9",
    "six==1.16.0",
    "sniffio==1.3.0",
    "sqlalchemy==2.0.23",
    "starlette==0.27.0",
    "typing-extensions==4.8.0",
    "uvicorn==0.24.0.post1",
    "watchfiles==0.21.0",
    "websockets==12.0",
]
